Print: add print(bool) and println(bool)
The lack of specialized Print::print(bool) method makes printing bools
more involved than they should be: the bool value is first implicitly
converted to an int to call print(int), which passes it on to
print(long), which then passes it to printNumber(), which finally
prints it in base 10.

Add specialized print(bool) and println(bool) methods to the Print
class to avoid all this hassle, and just write() '1' or '0' directly.

This change reduces the time needed for print((bool) 1) from 64usec to
8usec, and the size of the following little program from 1934 bytes to
1836 bytes.

  void setup()
  {
    Serial.begin(115200);
    Serial.println((long) 42);
    Serial.println((bool) 1);
  }
  void loop() {}